# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend service for the Cryptocash Battery on the TON network.",
        "title": "Battery implementation for TON network",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/admin/killswitch": {
            "get": {
                "security": [
                    {
                        "AdminToken": []
                    }
                ],
                "description": "Returns whether the battery is paused. Requires the admin token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get the battery kill switch (admin)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.KillswitchResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AdminToken": []
                    }
                ],
                "description": "Pauses or resumes all sends and prints. Requires the admin token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Set the battery kill switch (admin)",
                "parameters": [
                    {
                        "description": "paused flag",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.killswitchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.KillswitchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/users/{id}": {
            "get": {
                "security": [
                    {
                        "AdminToken": []
                    }
                ],
                "description": "Returns balance, reserved and available charges for a user. Requires the admin token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get a user's charges balance (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.BalanceResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/users/{id}/credit": {
            "post": {
                "security": [
                    {
                        "AdminToken": []
                    }
                ],
                "description": "Mock top-up: adds charges to a user's balance. Requires the admin token. Idempotent per idempotencyKey.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Credit charges to a user (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "amount + idempotency key",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.creditRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.AdminCreditResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/users/{id}/reconcile": {
            "get": {
                "security": [
                    {
                        "AdminToken": []
                    }
                ],
                "description": "Recomputes balance and reserved from the ledger entries and compares to the account. Requires the admin token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Reconcile a user's ledger (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ReconcileResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the authenticated user's id and public key",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.MeResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/balance": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the authenticated user's balance, reserved and available charges",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "charges"
                ],
                "summary": "Charges balance",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.BalanceResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/transactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the authenticated user's ledger entries, newest first",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "charges"
                ],
                "summary": "Charges history",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.TransactionsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.AdminCreditResponse": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "integer"
                },
                "balance": {
                    "type": "integer"
                },
                "entryId": {
                    "type": "string"
                },
                "idempotent": {
                    "type": "boolean"
                },
                "reserved": {
                    "type": "integer"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "handler.BalanceResponse": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "integer"
                },
                "balance": {
                    "type": "integer"
                },
                "reserved": {
                    "type": "integer"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "handler.EntryDTO": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "balanceAfter": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "op": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "reservedAfter": {
                    "type": "integer"
                }
            }
        },
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handler.KillswitchResponse": {
            "type": "object",
            "properties": {
                "paused": {
                    "type": "boolean"
                }
            }
        },
        "handler.MeResponse": {
            "type": "object",
            "properties": {
                "publicKey": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "handler.ReconcileResponse": {
            "type": "object",
            "properties": {
                "accountBalance": {
                    "type": "integer"
                },
                "accountReserved": {
                    "type": "integer"
                },
                "computedBalance": {
                    "type": "integer"
                },
                "computedReserved": {
                    "type": "integer"
                },
                "consistent": {
                    "type": "boolean"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "handler.TransactionsResponse": {
            "type": "object",
            "properties": {
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.EntryDTO"
                    }
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "handler.creditRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "idempotencyKey": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "handler.killswitchRequest": {
            "type": "object",
            "properties": {
                "paused": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminToken": {
            "type": "apiKey",
            "name": "X-Admin-Token",
            "in": "header"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}